Cleanup related to gyro setup and autonomous mode initializations
